Extract remaining helpers into modules
